pyproject.toml
[project]
name = "smart-parking-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-bcrypt>=1.0.1",
    "flask-cors>=6.0.0",
    "flask-login>=0.6.3",
    "flask-mail>=0.10.0",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "geopy>=2.4.1",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "qrcode[pil]>=8.2",
    "requests>=2.32.4",
]
